Change Styles and Paths

+++ src/components/guides/GuideBooking.jsx
@@ -1,30 +1,63 @@
 import React, { useEffect, useState } from "react";
-import { Link, useParams } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchGuides } from "../../redux/store/guidesSlice";
 import img from "../../assets/images/guide.jpg";
 import { fetchAvailableGuides } from "../../redux/store/availableGuideSlice";
-import avatar from "../../assets/images/avatar/avatar.png";
-import DatePicker from "react-datepicker";
+import Swal from "sweetalert2";
 
 const GuideDetails = () => {
   const { availableGuides } = useSelector((state) => state.availableGuides);
   const dispatch = useDispatch();
   const [showModal, setShowModal] = useState(false);
+  const navigate = useNavigate();
 
   useEffect(() => {
     dispatch(fetchAvailableGuides());
   }, [dispatch]);
 
   function onclick(event, data) {
-    console.log(data);
-    localStorage.setItem("guide", data);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "info",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Confirm Booking!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        console.log(data);
+        localStorage.setItem("guide", data);
+        navigate("/booking/driver");
+      }
+    });
   }
 
   return (
     <div>
       <div className="absolute top-0 h-20 w-full bg-black">
       </div>
+
+      <nav aria-label="breadcrumb" className="bg-custom-light px-4 pt-24 rounded">
+        <ol className="breadcrumb mb-0">
+          <li className="breadcrumb-item">
+            <Link
+              className="text-decoration-none link-secondary"
+              to="/vehicle"
+              replace
+            >
+              Booking
+            </Link>
+          </li>
+          <li className="breadcrumb-item active" aria-current="page">
+            Drivers
+          </li>
+          <li className="breadcrumb-item active" aria-current="page">
+            Vehicles
+          </li>
+        </ol>
+      </nav>
+
       {showModal && (
         <div className="shadow-lg">
           <div className="fixed inset-0 bg-black bg-opacity-25 backdrop-blur-sm flex items-center justify-center">
@@ -48,13 +81,10 @@ const GuideDetails = () => {
                 {/* Modal body */}
                 <div className="p-6 space-y-6">
                   <p className="text-base leading-relaxed text-gray-500 dark:text-gray-400">
-                    With less than a month to go before the European Union enacts new consumer privacy laws for its
-                    citizens, companies around the world are updating their terms of service agreements to comply.
-                  </p>
-                  <p className="text-base leading-relaxed text-gray-500 dark:text-gray-400">
-                    The European Union’s General Data Protection Regulation (G.D.P.R.) goes into effect on May 25 and is
-                    meant to ensure a common set of data rights in the European Union. It requires organizations to
-                    notify users as soon as possible of high-risk data breaches that could personally affect them.
+                    When you are select a guide for your tour make sure to check her/his price, rating, and language
+                    that guide fluent with.
+                    After selecting a guide according to your needs, click Select Guide button to Confirm your
+                    Selection.
                   </p>
                 </div>
                 {/* Modal footer */}
@@ -63,7 +93,7 @@ const GuideDetails = () => {
           </div>
         </div>
       )}
-      <div className="pt-24 px-2">
+      <div className="pt-4 px-3">
         <>
           <div
             id="alert-additional-content-5"
@@ -87,16 +117,11 @@ const GuideDetails = () => {
                 This is a dark alert
               </h3>
             </div>
-            <div className="mt-2 mb-4 text-sm text-gray-700 dark:text-gray-300">
-              More info about this info dark goes here. This example text is going to
-              run a bit longer so that you can see how spacing within an alert works
-              with this kind of content.
-            </div>
             <div className="flex">
               <button
                 type="button"
                 onClick={() => setShowModal(true)}
-                className="text-white bg-gray-700 hover:bg-gray-800 focus:ring-4 focus:outline-none focus:ring-gray-300 font-medium rounded-lg text-xs px-3 py-1.5 mr-2 text-center inline-flex items-center dark:bg-gray-600 dark:hover:bg-gray-500 dark:focus:ring-gray-600"
+                className="text-white mt-3 bg-gray-700 hover:bg-gray-800 focus:ring-4 focus:outline-none focus:ring-gray-300 font-medium rounded-lg text-xs px-3 py-1.5 mr-2 text-center inline-flex items-center dark:bg-gray-600 dark:hover:bg-gray-500 dark:focus:ring-gray-600"
               >
                 <svg
                   className="-ml-0.5 mr-2 h-4 w-4"
@@ -116,47 +141,56 @@ const GuideDetails = () => {
             </div>
           </div>
         </>
-
-
       </div>
-      <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-4 py-12">
+
+
+      <div className="mx-auto px-4 sm:px-6 lg:px-4 py-12">
         <div className="text-center pb-12">
+          <div className="flex mb-4">
+            <div className="w-1/2 h-12">
+              <h1 className="text-center text-3xl font-bold"></h1>
+            </div>
+            <div className="w-1/2 h-12">
+              <div className="absolute right-4"><Link to="/booking/driver">
+                <button
+                  className="flex rounded-md bg-blue-700 py-2 px-4 text-white transition-all duration-150 ease-in-out hover:bg-blue-600">
+                  SKIP GUIDES BOOKING & CONTINUE YOUR BOOKING <i className="bi bi-arrow-right ml-2"></i>
+                </button>
+              </Link></div>
+            </div>
+          </div>
+          <h1 className="text-3xl font-bold">Guide Booking</h1>
         </div>
         <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-2 lg:grid-cols-3 gap-6">
           {availableGuides?.length !== 0 && (
             availableGuides?.map((guide) => (
               <div className="no-underline cursor-pointer" key={guide.userId}>
-                <Link
-                  to={`/guides/${guide.userId}`}
-                  className="w-full bg-white rounded-lg sahdow-lg overflow-hidden flex flex-col justify-center items-center">
-                  <div>
-                    <img className="object-center object-cover w-[350px] h-[300px]"
-                         src={guide.userPhotoUrl ? guide.userPhotoUrl : img}
-                         alt="photo" />
+                <div className="flex justify-center h-[250px] ">
+                  <div className="flex flex-col md:flex-row md:max-w-xl rounded-lg bg-white shadow-lg">
+                    <Link to={`/guides/${guide.userId}`}>
+                      <img
+                        className=" w-full h-96 md:h-auto object-cover md:w-48 rounded-t-lg md:rounded-none md:rounded-l-lg"
+                        src={guide.userPhotoUrl ? guide.userPhotoUrl : img} alt="" />
+                    </Link>
+                    <div className="p-6 flex flex-col justify-start">
+                      <h5 className="text-gray-900 text-xl font-medium mb-2">Name - {guide.name}</h5>
+                      <p className="text-gray-700 text-base mb-4">
+                        Email - {guide.email}
+                      </p>
+                      <p className="text-gray-600 text-xs mb-4">Rating - {guide.rating}</p>
+                      <button
+                        className="flex rounded-md bg-black py-2 px-4 text-white transition-all duration-150 ease-in-out hover:bg-blue-600 text-center justify-center"
+                        onClick={event => onclick(event, guide.userId)}>
+                        BOOK GUIDE
+                      </button>
+                    </div>
                   </div>
-                  <div className="text-center py-8 sm:py-6">
-                    <p className="text-xl text-gray-700 font-bold mb-2">Name - {guide.name}</p>
-                    <p className="text-base text-gray-400 font-normal">Email - {guide.email}</p>
-                    <p className="text-base text-gray-400 font-normal">Rating - {guide.rating}</p>
-                  </div>
-                  <Link to="/booking/driver">
-                    <button
-                      className="flex rounded-md bg-blue-500 py-2 px-4 text-white transition-all duration-150 ease-in-out hover:bg-blue-600"
-                      onClick={event => onclick(event, guide.userId)}>
-                      SELECT
-                    </button>
-                  </Link>
-                </Link>
+                </div>
+
               </div>
             ))
           )}
         </div>
-        <Link to="/booking/driver">
-          <button
-            className="flex rounded-md bg-blue-500 py-2 px-4 text-white transition-all duration-150 ease-in-out hover:bg-blue-600">
-            NEXT
-          </button>
-        </Link>
       </div>
     </div>
   );
